+++ vite.config.ts
@@ -225,7 +225,7 @@ export default defineConfig(({ mode }) => {
           cleanupOutdatedCaches: true,
           clientsClaim: true,
           skipWaiting: true,
-          globPatterns: ["**/*.{js,css,html,ico,png,svg,woff,woff2,json,nmea}"],
+          globPatterns: ["**/*.{js,css,html,ico,png,svg,woff,woff2,json,nmea,wasm}"],
           globIgnores: ["**/tracks.zip"],
           navigateFallbackDenylist: [/^\/~oauth/],
           runtimeCaching: [